API update

+++ be-live-admin/cmd/admin/handler/user_handler.go
@@ -3,15 +3,16 @@ package handler
 import (
 	"errors"
 	"fmt"
-	"gitlab/live/be-live-api/conf"
-	"gitlab/live/be-live-api/dto"
-	"gitlab/live/be-live-api/model"
-	"gitlab/live/be-live-api/service"
-	"gitlab/live/be-live-api/utils"
+	"gitlab/live/be-live-admin/conf"
+	"gitlab/live/be-live-admin/dto"
+	"gitlab/live/be-live-admin/model"
+	"gitlab/live/be-live-admin/service"
+	"gitlab/live/be-live-admin/utils"
 	"io"
 	"log"
 	"net/http"
 	"os"
+	"slices"
 	"strconv"
 	"strings"
 
@@ -49,6 +50,7 @@ func (h *userHandler) register() {
 	group.GET("", h.page)
 	group.POST("", h.createUser)
 	group.PUT("/:id", h.updateUser)
+	group.PATCH("/:id/change-password", h.changePassword)
 	group.GET("/:id", h.byId)
 	group.DELETE("/:id", h.deleteByID)
 
@@ -76,6 +78,22 @@ func (h *userHandler) deleteByID(c echo.Context) error {
 	}
 
 	currentUser := c.Get("user").(*utils.Claims)
+	deletedUser, err := h.srv.User.FindByID(uint(id))
+	if err != nil {
+
+	}
+
+	if deletedUser == nil {
+		return utils.BuildErrorResponse(c, http.StatusNotFound, errors.New("not found"), nil)
+	}
+
+	// remove avatar
+	if deletedUser.AvatarFileName.Valid {
+
+		avatarPath := fmt.Sprintf("%s%s", h.avatarFolder, deletedUser.AvatarFileName.String)
+		avatarsToRemove := []string{avatarPath}
+		go utils.RemoveFiles(avatarsToRemove)
+	}
 
 	if err := h.srv.User.DeleteByID(uint(id), currentUser.ID); err != nil {
 		return utils.BuildErrorResponse(c, http.StatusInternalServerError, err, nil)
@@ -114,6 +132,10 @@ func (h *userHandler) createUser(c echo.Context) error {
 
 		if !isImage {
 			return utils.BuildErrorResponse(c, http.StatusBadRequest, errors.New("file is not an image"), nil)
+		}
+
+		if file.Size > utils.MAX_IMAGE_SIZE {
+			return utils.BuildErrorResponse(c, http.StatusBadRequest, nil, "Image size exceeds the 1MB limit")
 		}
 
 		// save avatar
@@ -175,11 +197,85 @@ func (h *userHandler) updateUser(c echo.Context) error {
 
 	currentUser := c.Get("user").(*utils.Claims)
 	req.UpdatedByID = &currentUser.ID
-	data, err := h.srv.User.UpdateUser(&req, uint(id))
+
+	targetUser, err := h.srv.User.FindByID(uint(id))
+	if err != nil {
+		return utils.BuildErrorResponse(c, http.StatusInternalServerError, err, nil)
+	}
+	if targetUser == nil {
+		return utils.BuildErrorResponse(c, http.StatusNotFound, errors.New("not found"), nil)
+	}
+
+	var data *dto.UpdateUserResponse
+	// validate super admin user
+	if currentUser.RoleType == model.SUPPERADMINROLE {
+		if currentUser.ID == uint(id) {
+			return utils.BuildErrorResponse(c, http.StatusBadRequest, errors.New("invalid request, super admin don't update itself"), nil)
+		}
+	} else {
+		// validate admin user
+		if currentUser.ID != uint(id) {
+			return utils.BuildErrorResponse(c, http.StatusBadRequest, errors.New("invalid request, not allow update info of other account"), nil)
+		}
+	}
+
+	data, err = h.srv.User.UpdateUser(&req, uint(id))
 	if err != nil {
 		return utils.BuildErrorResponse(c, http.StatusInternalServerError, err, nil)
 	}
 	adminLog := h.srv.Admin.MakeAdminLogModel(uint(id), model.UpdateUserAction, fmt.Sprintf(" %s update_user request", currentUser.Email))
+
+	err = h.srv.Admin.CreateLog(adminLog)
+
+	if err != nil {
+		return c.JSON(http.StatusInternalServerError, map[string]string{"error": "Failed to created admin log"})
+	}
+
+	return utils.BuildSuccessResponseWithData(c, http.StatusOK, data)
+}
+
+func (h *userHandler) changePassword(c echo.Context) error {
+	var err error
+	var id int
+	var req dto.ChangePasswordRequest
+	if err := utils.BindAndValidate(c, &req); err != nil {
+		return utils.BuildErrorResponse(c, http.StatusBadRequest, err, nil)
+	}
+
+	if c.Param("id") != "" {
+		id, err = strconv.Atoi(c.Param("id"))
+		if err != nil {
+			return utils.BuildErrorResponse(c, http.StatusBadRequest, errors.New("invalid id parameter"), nil)
+		}
+	}
+
+	if req.Password != req.ConfirmPassword {
+		return utils.BuildErrorResponse(c, http.StatusBadRequest, errors.New("password not match confirm_password"), nil)
+	}
+
+	currentUser := c.Get("user").(*utils.Claims)
+
+	targetUser, err := h.srv.User.FindByID(uint(id))
+	if err != nil {
+		return utils.BuildErrorResponse(c, http.StatusInternalServerError, err, nil)
+	}
+	if targetUser == nil {
+		return utils.BuildErrorResponse(c, http.StatusNotFound, errors.New("not found"), nil)
+	}
+
+	var data *dto.UpdateUserResponse
+	// validate admin user
+	if currentUser.RoleType == model.ADMINROLE {
+		if !slices.Contains([]model.RoleType{model.STREAMER, model.USERROLE}, targetUser.Role.Type) && currentUser.ID != uint(id) {
+			return utils.BuildErrorResponse(c, http.StatusBadRequest, errors.New("invalid request, just allow change password self and streamer, user"), nil)
+		}
+	}
+
+	data, err = h.srv.User.ChangePassword(targetUser, &req, uint(id), currentUser.ID)
+	if err != nil {
+		return utils.BuildErrorResponse(c, http.StatusInternalServerError, err, nil)
+	}
+	adminLog := h.srv.Admin.MakeAdminLogModel(currentUser.ID, model.ChangeUserPasswordAction, fmt.Sprintf(" %s change_user_password request", currentUser.Email))
 
 	err = h.srv.Admin.CreateLog(adminLog)
